Separated logging utils. Added drip-feed logging for command-less usage.
