+++ src/cli/commands/clean/index.js
@@ -3,7 +3,7 @@ const del = require('del');
 
 // Ours
 const {packs, unpack} = require('../../../utils/async');
-const {log} = require('../../../utils');
+const {log} = require('../../../utils/console');
 const {command} = require('../');
 const {OPTIONS, DEFAULTS, USAGE, MESSAGES} = require('./constants');
