+++ src/cli/commands/serve-basic-story/index.js
@@ -10,7 +10,7 @@ const serveStatic = require('serve-static');
 
 // Ours
 const {pack, throws} = require('../../../utils/async');
-const {log} = require('../../../utils');
+const {log} = require('../../../utils/console');
 const {
   BUILD_DIR, D_KEY, DEFAULTS, KEY, TASK_NAMES
 } = require('../build-basic-story/constants');
