+++ src/projects.js
@@ -13,8 +13,9 @@ const ourPkg = require('../package');
 const {CONFIG_FILE_NAME} = require('./constants');
 const {hvy, ok} = require('./string-styles');
 const {packs, prequire, unpack} = require('./utils/async');
+const {log} = require('./utils/console');
 const {indented} = require('./utils/strings');
-const {identity, log, pretty} = require('./utils');
+const {identity, pretty} = require('./utils');
 
 // Wrapped
 const clone = packs(pify(copyTemplateDir));
