+++ src/utils/async.js
@@ -5,6 +5,10 @@ const pump = require('pump');
 // Wrapped
 const pumpAsync = pify(pump);
 
+function delay(ms) {
+  return new Promise(resolve => setTimeout(() => resolve(), ms));
+}
+
 function pack(promise) {
   return promise
   .then(result => [null, result])
@@ -45,6 +49,7 @@ function pumped(...streams) {
 }
 
 module.exports = {
+  delay,
   pack,
   packs,
   throws,
